fix(environments) Disallow `/` in future environment names
While we can't reasonably change existing data, we can prevent more in
the future. Having `/` in an environment name makes it impossible to
hide the environment.

Event ingestion applies the new rules as well. This results in
environment being blanked, and event.errors being populated with
validation errors. This replicates the behavior that releases have.

Refs APP-323
Refs ISSUE-24